# swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This package provides Swagger API documentation for the Iznik API.\nThe problem is that we need to explicitly define the API operations here for them to be detected.",
    "title": "Iznik API.",
    "termsOfService": "https://www.ilovefreegle.org/terms",
    "contact": {
      "name": "Freegle Geeks",
      "url": "https://www.ilovefreegle.org/help",
      "email": "geeks@ilovefreegle.org"
    },
    "license": {
      "name": "GPL v2",
      "url": "https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html"
    },
    "version": "1.0"
  },
  "host": "api.ilovefreegle.org",
  "basePath": "/api",
  "paths": {
...
    "/address": {
      "get": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "# Returns all addresses for the authenticated user",
        "tags": [
          "address"
        ],
        "summary": "List addresses for user",
        "operationId": "listAddresses",
        "responses": {
          "200": {
            "$ref": "#/responses/addressesResponse"
          }
        }
      }
    },
    "/address/{id}": {
      "get": {
        "description": "Returns a single address by ID",
        "tags": [
          "address"
        ],
        "summary": "Get address by ID",
        "operationId": "getAddress",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Address ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/addressResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/chat": {
      "get": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "# Returns all chats for the authenticated user",
        "tags": [
          "chat"
        ],
        "summary": "List chats for user",
        "operationId": "listChats",
        "responses": {
          "200": {
            "$ref": "#/responses/chatsResponse"
          }
        }
      }
    },
    "/chat/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Returns a single chat by ID",
        "tags": [
          "chat"
        ],
        "summary": "Get chat by ID",
        "operationId": "getChat",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Chat ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/chatResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/message/{ids}": {
      "get": {
        "description": "Returns messages by ID (comma separated)",
        "tags": [
          "message"
        ],
        "summary": "Get messages by ID",
        "operationId": "getMessages",
        "parameters": [
          {
            "type": "string",
            "description": "Message IDs (comma separated)",
            "name": "ids",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/messagesResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Returns a single user by ID, or the current user if no ID",
        "tags": [
          "user"
        ],
        "summary": "Get user by ID",
        "operationId": "getUser",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/userResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/config/{key}": {
      "get": {
        "description": "Returns configuration by key",
        "tags": [
          "config"
        ],
        "summary": "Get configuration",
        "operationId": "getConfig",
        "parameters": [
          {
            "type": "string",
            "description": "Configuration key",
            "name": "key",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/configResponse"
          }
        }
      }
    },
    "/config/admin/spam_keywords": {
      "get": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Returns all spam keywords (Support/Admin only)",
        "tags": [
          "config"
        ],
        "summary": "List spam keywords",
        "operationId": "listSpamKeywords",
        "responses": {
          "200": {
            "$ref": "#/responses/spamKeywordsResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Creates a new spam keyword (Support/Admin only)",
        "tags": [
          "config"
        ],
        "summary": "Create spam keyword",
        "operationId": "createSpamKeyword",
        "parameters": [
          {
            "description": "Spam keyword object",
            "name": "spam_keyword",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateSpamKeywordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/spamKeywordResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/config/admin/spam_keywords/{id}": {
      "delete": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Deletes a spam keyword by ID (Support/Admin only)",
        "tags": [
          "config"
        ],
        "summary": "Delete spam keyword",
        "operationId": "deleteSpamKeyword",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Spam keyword ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/successResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/errorResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/config/admin/worry_words": {
      "get": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Returns all worry words (Support/Admin only)",
        "tags": [
          "config"
        ],
        "summary": "List worry words",
        "operationId": "listWorryWords",
        "responses": {
          "200": {
            "$ref": "#/responses/worryWordsResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Creates a new worry word (Support/Admin only)",
        "tags": [
          "config"
        ],
        "summary": "Create worry word",
        "operationId": "createWorryWord",
        "parameters": [
          {
            "description": "Worry word object",
            "name": "worry_word",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateWorryWordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/worryWordResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/config/admin/worry_words/{id}": {
      "delete": {
        "security": [
          {
            "BearerAuth": [
              "[]"
            ]
          }
        ],
        "description": "Deletes a worry word by ID (Support/Admin only)",
        "tags": [
          "config"
        ],
        "summary": "Delete worry word",
        "operationId": "deleteWorryWord",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Worry word ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/successResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/errorResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ChatRoom": {
      "type": "object",
      "properties": {
        "chattype": {
          "type": "string",
          "x-go-name": "Chattype"
        },
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID"
        },
        "user1": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "User1"
        },
        "user2": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "User2"
        }
      },
      "x-go-package": "github.com/freegle/iznik-server-go/chat"
    },
...
    "Map": {
      "description": "Map is a shortcut for map[string]interface{}, useful for JSON returns",
      "type": "object",
      "additionalProperties": {},
      "x-go-package": "github.com/gofiber/fiber/v2"
    },
...
    "ConfigItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID"
        },
        "key": {
          "type": "string",
          "x-go-name": "Key"
        },
        "value": {
          "type": "string",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/freegle/iznik-server-go/config"
    },
...
    "CreateSpamKeywordRequest": {
      "type": "object",
      "properties": {
        "word": {
          "type": "string",
          "x-go-name": "Word"
        },
        "exclude": {
          "type": "string",
          "x-go-name": "Exclude"
        },
        "action": {
          "type": "string",
          "x-go-name": "Action"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/freegle/iznik-server-go/config"
    },
    "CreateWorryWordRequest": {
      "type": "object",
      "properties": {
        "keyword": {
          "type": "string",
          "x-go-name": "Keyword"
        },
        "substance": {
          "type": "string",
          "x-go-name": "Substance"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/freegle/iznik-server-go/config"
    }
  },
  "responses": {
...
    "addressResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Address"
      }
    },
    "addressesResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Address"
        }
      }
    },
    "chatResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ChatRoom"
      }
    },
    "chatsResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/ChatRoom"
        }
      }
    },
    "errorResponse": {
      "description": "errorResponse is the error response",
      "schema": {
        "$ref": "#/definitions/Map"
      }
    },
    "messagesResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Message"
        }
      }
    },
    "userResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/User"
      }
    },
    "configResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ConfigItem"
      }
    },
    "spamKeywordsResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/SpamKeyword"
        }
      }
    },
    "spamKeywordResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/SpamKeyword"
      }
    },
    "worryWordsResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/WorryWord"
        }
      }
    },
    "worryWordResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/WorryWord"
      }
    },
    "successResponse": {
      "description": "Success",
      "schema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}